# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Doomy Vision API
  description: >
    API aislada que conecta el Bridge (Ray-Ban Meta / fallback de teléfono) con Doomy.
    Refleja solo endpoints que existen realmente en backend/src/routes/ — no aspiracional.
    Ver docs/DOOMY_VISION_BRIDGE_API.md para notas de uso más detalladas.
  version: "1.0.0"
servers:
  - url: /api/doomy-vision/v1
    description: Backend Doomy Vision (local, o el dominio de Railway una vez desplegado)
security:
  - bearerAuth: []
paths:
...
  /session:
    post:
      summary: Crear una sesión
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device_type:
                  type: string
                mode:
                  type: string
                  enum: [real, mock, phone]
      responses:
        "201":
          description: Sesión creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /session/{id}:
    get:
      summary: Consultar el estado de una sesión
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Estado actual de la sesión
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /session/{id}/reset:
    post:
      summary: Limpiar historial, imagen activa y resumen visual de una sesión
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Sesión reseteada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error estructurado (ver docs/DOOMY_VISION_BRIDGE_API.md para la tabla completa de códigos)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Session:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        device_id:
          type: string
        device_type:
          type: string
        mode:
          type: string
        created_at:
          type: string
          format: date-time
        turns:
          type: number
        has_active_image:
          type: boolean
        active_image_summary:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            dv_code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true